+++ Game/Game/Menus/Concrete Menus/Score Screens/ScoreScreen.cpp
@@ -369,7 +369,7 @@ bool ScoreScreen::UseZoomIn = true;
 		{
 			SlideIn( 0 );
 			std::vector<Vector2> copy = std::vector<Vector2>( DrawPile::BubbleScale.size() );
-			for ( int i = 0; i < DrawPile::BubbleScale.size(); ++i )
+			for ( int i = 0; i < static_cast<int>( DrawPile::BubbleScale.size() ); ++i )
 			{
 				copy[ i ] = ( DrawPile::BubbleScale[ i ] - Vector2(1) ) *.3f + Vector2(1);
 			}
